feat(pre-api): add mgmt account creds, don't purge (#19)
* feat(pre-api): add mgmt account creds, don't purge

* fix(creds): remove tabs

* fix(creds): spacing

* fix(creds): spacing

+++ sources/scripts/creds.sh
@@ -9,6 +9,7 @@ AFT=false
 CT=false
 AUDIT=false
 LOG_ARCHIVE=false
+NO_PURGE=
 
 #Ensure at option was specified
 if [ $# -eq 0 ]; then
@@ -20,46 +21,51 @@ fi
 #Parse options and set flags
 while [ ! $# -eq 0 ]
 do
-	case "$1" in
-        --account-id)
-            USER_DEFINED_ACCOUNT=true
-            USER_DEFINED_ACCOUNT_ID=$2
-            echo $ACCOUNT_ID
-            ;;
-		--aft-mgmt)
-			AFT=true
-			;;
-		--ct-mgmt)
-			CT=true
-			;;
-		--ct-audit)
-			AUDIT=true
-			;;
-		--ct-log-archive)
-			LOG_ARCHIVE=true
-			;;
-        --help)
-            echo ""
-            echo "creds.sh creates an AWS CLI credential file leveraging AWSAFTExecutionRole for specified accounts"
-            echo ""
-            echo "** creds.sh should be run from the AFT Management account with a role that can assume aws-aft-AdministratorRole **"
-            echo ""
-            echo "usage: creds.sh [--account account_id] [--ct-mgmt] [--aft-mgmt]"
-            echo ""
-            echo "--account-id - Create a default credential profile for the given account number.   Profile name: default"
-            echo "--aft-mgmt   - Create a credential profile for AFT Management account.             Profile name: aft-management"
-			      echo "--ct-mgmt    - Create a credential profile for Control Tower Management account.   Profile name: ct-management"
-			      echo "--audit      - Create a credential profile for Control Tower Audit account.        Profile name: ct-audit"
-			      echo "--log        - Create a credential profile for Control Tower Log Archive account.  Profile name: ct-log-archive"
-            exit
-			;;
-	esac
-	shift
+  case "$1" in
+    --account-id)
+        USER_DEFINED_ACCOUNT=true
+        USER_DEFINED_ACCOUNT_ID=$2
+        echo $ACCOUNT_ID
+        ;;
+    --aft-mgmt)
+        AFT=true
+        ;;
+    --ct-mgmt)
+        CT=true
+        ;;
+    --ct-audit)
+        AUDIT=true
+        ;;
+    --ct-log-archive)
+        LOG_ARCHIVE=true
+        ;;
+    --no-purge)
+        NO_PURGE=true
+        ;;
+    --help)
+        echo ""
+        echo "creds.sh creates an AWS CLI credential file leveraging AWSAFTExecutionRole for specified accounts"
+        echo ""
+        echo "** creds.sh should be run from the AFT Management account with a role that can assume aws-aft-AdministratorRole **"
+        echo ""
+        echo "usage: creds.sh [--account account_id] [--ct-mgmt] [--aft-mgmt]"
+        echo ""
+        echo "--account-id - Create a default credential profile for the given account number.   Profile name: default"
+        echo "--aft-mgmt   - Create a credential profile for AFT Management account.             Profile name: aft-management"
+        echo "--ct-mgmt    - Create a credential profile for Control Tower Management account.   Profile name: ct-management"
+        echo "--audit      - Create a credential profile for Control Tower Audit account.        Profile name: ct-audit"
+        echo "--log        - Create a credential profile for Control Tower Log Archive account.  Profile name: ct-log-archive"
+        exit
+        ;;
+  esac
+  shift
 done
 
 # Remove Credentials file, if exists
 mkdir -p ~/.aws
-rm -f ~/.aws/credentials
+if [ -z "$NO_PURGE" ]; then
+  rm -f ~/.aws/credentials
+fi
 
 #Lookup SSM Parameters
 AFT_MGMT_ROLE=$(aws ssm get-parameter --name /aft/resources/iam/aft-administrator-role-name | jq --raw-output ".Parameter.Value")
